+++ .circleci/test-server.sh
@@ -198,7 +198,6 @@ if [ -n ${HASURA_GRAPHQL_DATABASE_URL_2:-} ] ; then
 fi
 export WEBHOOK_FROM_ENV="http://127.0.0.1:5592"
 export HASURA_GRAPHQL_STRINGIFY_NUMERIC_TYPES=true
-export HASURA_GRAPHQL_ADMIN_INTERNAL_ERRORS=true
 
 HGE_PIDS=""
 WH_PID=""
