+++ server/src-lib/Hasura/GraphQL/Transport/HTTP.hs
@@ -11,7 +11,8 @@ import           Hasura.GraphQL.Logging
 import           Hasura.GraphQL.Transport.HTTP.Protocol
 import           Hasura.Prelude
 import           Hasura.RQL.Types
-import           Hasura.Server.Context
+import           Hasura.Server.API.Types
+import           Hasura.Server.Init.Config
 import           Hasura.Server.Utils                    (RequestId)
 import           Hasura.Server.Version                  (HasVersion)
 
@@ -61,12 +62,12 @@ runGQBatched
      , MonadReader E.ExecutionCtx m
      )
   => RequestId
-  -> GraphQLResponseConfig
+  -> ResponseInternalErrorsConfig
   -> UserInfo
   -> [N.Header]
   -> GQLBatchedReqs GQLQueryText
   -> m (HttpResponse EncJSON)
-runGQBatched reqId responseConfig userInfo reqHdrs reqs =
+runGQBatched reqId responseErrorsConfig userInfo reqHdrs reqs =
   case reqs of
     GQLSingleRequest req ->
       runGQ reqId userInfo reqHdrs req
@@ -74,14 +75,13 @@ runGQBatched reqId responseConfig userInfo reqHdrs reqs =
       -- It's unclear what we should do if we receive multiple
       -- responses with distinct headers, so just do the simplest thing
       -- in this case, and don't forward any.
-      let removeHeaders =
+      let includeInternal = shouldIncludeInternal (userRole userInfo) responseErrorsConfig
+          removeHeaders =
             flip HttpResponse []
             . encJFromList
-            . map (either (encJFromJValue . encodeGQErr responseConfig isAdminRole) _hrBody)
+            . map (either (encJFromJValue . encodeGQErr includeInternal) _hrBody)
           try = flip catchError (pure . Left) . fmap Right
       removeHeaders <$> traverse (try . runGQ reqId userInfo reqHdrs) batch
-  where
-    isAdminRole = isAdmin $ userRole userInfo
 
 runHasuraGQ
   :: ( MonadIO m
