+++ server/src-lib/Hasura/RQL/Types/Error.hs
@@ -46,7 +46,6 @@ module Hasura.RQL.Types.Error
        ) where
 
 import           Hasura.Prelude
-import           Hasura.Server.Context
 
 import           Control.Arrow.Extended
 import           Data.Aeson
@@ -173,8 +172,8 @@ noInternalQErrEnc (QErr jPath _ msg code _) =
   , "code"  .= show code
   ]
 
-encodeGQLErr :: GraphQLResponseConfig -> Bool -> QErr -> Value
-encodeGQLErr responseConfig isAdminRole (QErr jPath _ msg code mIE) =
+encodeGQLErr :: Bool -> QErr -> Value
+encodeGQLErr includeInternal (QErr jPath _ msg code mIE) =
   object
   [ "message" .= msg
   , "extensions" .= extnsObj
@@ -186,11 +185,6 @@ encodeGQLErr responseConfig isAdminRole (QErr jPath _ msg code mIE) =
                   , "path" .= encodeJSONPath jPath
                   ]
     internal = maybe [] (\ie -> ["internal" .= ie]) mIE
-
-    -- Include 'internal' if dev-mode is set to `true`
-    -- OR (current role is admin AND admin-internal-errors is set to `true`)
-    includeInternal = _gscDeveloperMode responseConfig
-                      || (_gscAdminInternalErrors responseConfig && isAdminRole)
 
 -- whether internal should be included or not
 encodeQErr :: Bool -> QErr -> Value
