+++ server/src-lib/Hasura/Server/App.hs
@@ -37,24 +37,24 @@ import qualified Text.Mustache                          as M
 import qualified Web.Spock.Core                         as Spock
 
 import           Hasura.EncJSON
+import           Hasura.GraphQL.Resolve.Action
 import           Hasura.Prelude                         hiding (get, put)
 import           Hasura.RQL.DDL.Schema
 import           Hasura.RQL.Types
 import           Hasura.RQL.Types.Run
+import           Hasura.Server.API.Config               (runGetConfig)
+import           Hasura.Server.API.Query
+import           Hasura.Server.API.Types
 import           Hasura.Server.Auth                     (AuthMode (..), UserAuthentication (..))
 import           Hasura.Server.Compression
-import           Hasura.Server.Config                   (runGetConfig)
-import           Hasura.Server.Context
 import           Hasura.Server.Cors
 import           Hasura.Server.Init
 import           Hasura.Server.Logging
 import           Hasura.Server.Middleware               (corsMiddleware)
 import           Hasura.Server.Migrate                  (migrateCatalog)
-import           Hasura.Server.Query
 import           Hasura.Server.Utils
 import           Hasura.Server.Version
 import           Hasura.SQL.Types
-import           Hasura.GraphQL.Resolve.Action
 
 import qualified Hasura.GraphQL.Execute                 as E
 import qualified Hasura.GraphQL.Execute.LiveQuery       as EL
@@ -63,7 +63,7 @@ import qualified Hasura.GraphQL.Transport.HTTP          as GH
 import qualified Hasura.GraphQL.Transport.HTTP.Protocol as GH
 import qualified Hasura.GraphQL.Transport.WebSocket     as WS
 import qualified Hasura.Logging                         as L
-import qualified Hasura.Server.PGDump                   as PGD
+import qualified Hasura.Server.API.PGDump               as PGD
 
 
 data SchemaCacheRef
@@ -90,20 +90,20 @@ data SchemaCacheRef
 
 data ServerCtx
   = ServerCtx
-  { scPGExecCtx             :: !PGExecCtx
-  , scConnInfo              :: !Q.ConnInfo
-  , scLogger                :: !(L.Logger L.Hasura)
-  , scCacheRef              :: !SchemaCacheRef
-  , scAuthMode              :: !AuthMode
-  , scManager               :: !HTTP.Manager
-  , scSQLGenCtx             :: !SQLGenCtx
-  , scEnabledAPIs           :: !(S.HashSet API)
-  , scInstanceId            :: !InstanceId
-  , scPlanCache             :: !E.PlanCache
-  , scLQState               :: !EL.LiveQueriesState
-  , scEnableAllowlist       :: !Bool
-  , scEkgStore              :: !EKG.Store
-  , scGraphqlResponseConfig :: !GraphQLResponseConfig
+  { scPGExecCtx                    :: !PGExecCtx
+  , scConnInfo                     :: !Q.ConnInfo
+  , scLogger                       :: !(L.Logger L.Hasura)
+  , scCacheRef                     :: !SchemaCacheRef
+  , scAuthMode                     :: !AuthMode
+  , scManager                      :: !HTTP.Manager
+  , scSQLGenCtx                    :: !SQLGenCtx
+  , scEnabledAPIs                  :: !(S.HashSet API)
+  , scInstanceId                   :: !InstanceId
+  , scPlanCache                    :: !E.PlanCache
+  , scLQState                      :: !EL.LiveQueriesState
+  , scEnableAllowlist              :: !Bool
+  , scEkgStore                     :: !EKG.Store
+  , scResponseInternalErrorsConfig :: !ResponseInternalErrorsConfig
   }
 
 data HandlerCtx
@@ -213,7 +213,7 @@ class MetadataApiAuthorization m where
 mkSpockAction
   :: (HasVersion, MonadIO m, FromJSON a, ToJSON a, UserAuthentication m, HttpLog m)
   => ServerCtx
-  -> (GraphQLResponseConfig -> Bool -> QErr -> Value)
+  -> (Bool -> QErr -> Value)
   -- ^ `QErr` JSON encoder function
   -> (QErr -> QErr)
   -- ^ `QErr` modifier
@@ -234,7 +234,8 @@ mkSpockAction serverCtx qErrEncoder qErrModifier apiHandler = do
                  return userInfoE
 
     let handlerState = HandlerCtx serverCtx userInfo headers requestId
-        curRole = userRole userInfo
+        includeInternal = shouldIncludeInternal (userRole userInfo) $
+                          scResponseInternalErrorsConfig serverCtx
 
     (serviceTime, (result, q)) <- withElapsedTime $ case apiHandler of
       AHGet handler -> do
@@ -242,7 +243,7 @@ mkSpockAction serverCtx qErrEncoder qErrModifier apiHandler = do
         return (res, Nothing)
       AHPost handler -> do
         parsedReqE <- runExceptT $ parseBody reqBody
-        parsedReq  <- either (logErrorAndResp (Just userInfo) requestId req (Left reqBody) (isAdmin curRole) headers . qErrModifier)
+        parsedReq  <- either (logErrorAndResp (Just userInfo) requestId req (Left reqBody) includeInternal headers . qErrModifier)
                       return parsedReqE
         res <- lift $ runReaderT (runExceptT $ handler parsedReq) handlerState
         return (res, Just parsedReq)
@@ -253,13 +254,11 @@ mkSpockAction serverCtx qErrEncoder qErrModifier apiHandler = do
     -- log and return result
     case modResult of
       Left err  -> let jErr = maybe (Left reqBody) (Right . toJSON) q
-                   in logErrorAndResp (Just userInfo) requestId req jErr (isAdmin curRole) headers err
+                   in logErrorAndResp (Just userInfo) requestId req jErr includeInternal headers err
       Right res -> logSuccessAndResp (Just userInfo) requestId req (fmap toJSON q) res (Just (ioWaitTime, serviceTime)) headers
 
     where
       logger = scLogger serverCtx
-      responseConfig = scGraphqlResponseConfig serverCtx
-
 
       logErrorAndResp
         :: (MonadIO m, HttpLog m)
@@ -271,10 +270,10 @@ mkSpockAction serverCtx qErrEncoder qErrModifier apiHandler = do
         -> [HTTP.Header]
         -> QErr
         -> Spock.ActionCtxT ctx m a
-      logErrorAndResp userInfo reqId req reqBody isAdminRole headers qErr = do
+      logErrorAndResp userInfo reqId req reqBody includeInternal headers qErr = do
         lift $ logHttpError logger userInfo reqId req reqBody qErr headers
         Spock.setStatus $ qeStatus qErr
-        Spock.json $ qErrEncoder responseConfig isAdminRole qErr
+        Spock.json $ qErrEncoder includeInternal qErr
 
       logSuccessAndResp userInfo reqId req reqBody result qTime reqHeaders =
         case result of
@@ -332,10 +331,10 @@ v1Alpha1GQHandler query = do
   enableAL  <- scEnableAllowlist . hcServerCtx <$> ask
   logger    <- scLogger . hcServerCtx <$> ask
   requestId <- asks hcRequestId
-  responseConfig <- scGraphqlResponseConfig . hcServerCtx <$> ask
+  responseErrorsConfig <- scResponseInternalErrorsConfig . hcServerCtx <$> ask
   let execCtx = E.ExecutionCtx logger sqlGenCtx pgExecCtx planCache
                 (lastBuiltSchemaCache sc) scVer manager enableAL
-  flip runReaderT execCtx $ GH.runGQBatched requestId responseConfig userInfo reqHeaders query
+  flip runReaderT execCtx $ GH.runGQBatched requestId responseErrorsConfig userInfo reqHeaders query
 
 v1GQHandler
   :: (HasVersion, MonadIO m)
@@ -471,10 +470,10 @@ mkWaiApp
   -> S.HashSet API
   -> EL.LiveQueriesOptions
   -> E.PlanCacheOptions
-  -> GraphQLResponseConfig
+  -> ResponseInternalErrorsConfig
   -> m HasuraApp
 mkWaiApp isoLevel logger sqlGenCtx enableAL pool ci httpManager mode corsCfg enableConsole consoleAssetsDir
-         enableTelemetry instanceId apis lqOpts planCacheOptions responseConfig = do
+         enableTelemetry instanceId apis lqOpts planCacheOptions responseErrorsConfig = do
 
     (planCache, schemaCacheRef, cacheBuiltTime) <- migrateAndInitialiseSchemaCache
     let getSchemaCache = first lastBuiltSchemaCache <$> readIORef (_scrCache schemaCacheRef)
@@ -501,7 +500,7 @@ mkWaiApp isoLevel logger sqlGenCtx enableAL pool ci httpManager mode corsCfg ena
                     , scLQState         =  lqState
                     , scEnableAllowlist =  enableAL
                     , scEkgStore        =  ekgStore
-                    , scGraphqlResponseConfig = responseConfig
+                    , scResponseInternalErrorsConfig = responseErrorsConfig
                     }
 
     when (isDeveloperAPIEnabled serverCtx) $ do
@@ -584,20 +583,20 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
 
       Spock.post "v1alpha1/graphql/explain" gqlExplainAction
 
-      Spock.post "v1/query" $ spockAction encodeQErr' id $
+      Spock.post "v1/query" $ spockAction encodeQErr id $
         mkPostHandler $ mkAPIRespHandler v1QueryHandler
 
       Spock.post ("api/1/table" <//> Spock.var <//> Spock.var) $ \tableName queryType ->
-        mkSpockAction serverCtx (const encodeQErr) id $ mkPostHandler $
+        mkSpockAction serverCtx encodeQErr id $ mkPostHandler $
           mkAPIRespHandler $ legacyQueryHandler (TableName tableName) queryType
 
     when enablePGDump $
 
-      Spock.post "v1alpha1/pg_dump" $ spockAction encodeQErr' id $
+      Spock.post "v1alpha1/pg_dump" $ spockAction encodeQErr id $
         mkPostHandler v1Alpha1PGDumpHandler
 
     when enableConfig $
-      Spock.get "v1alpha1/config" $ spockAction encodeQErr' id $
+      Spock.get "v1alpha1/config" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           let res = encJFromJValue $ runGetConfig
@@ -615,22 +614,22 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
         mkPostHandler $ mkAPIRespHandler v1GQHandler
 
     when (isDeveloperAPIEnabled serverCtx) $ do
-      Spock.get "dev/ekg" $ spockAction encodeQErr' id $
+      Spock.get "dev/ekg" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EKG.sampleAll $ scEkgStore serverCtx
           return $ JSONResp $ HttpResponse (encJFromJValue $ EKG.sampleToJson respJ) []
-      Spock.get "dev/plan_cache" $ spockAction encodeQErr' id $
+      Spock.get "dev/plan_cache" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ E.dumpPlanCache $ scPlanCache serverCtx
           return $ JSONResp $ HttpResponse (encJFromJValue respJ) []
-      Spock.get "dev/subscriptions" $ spockAction encodeQErr' id $
+      Spock.get "dev/subscriptions" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EL.dumpLiveQueriesState False $ scLQState serverCtx
           return $ JSONResp $ HttpResponse (encJFromJValue respJ) []
-      Spock.get "dev/subscriptions/extended" $ spockAction encodeQErr' id $
+      Spock.get "dev/subscriptions/extended" $ spockAction encodeQErr id $
         mkGetHandler $ do
           onlyAdmin
           respJ <- liftIO $ EL.dumpLiveQueriesState True $ scLQState serverCtx
@@ -644,11 +643,10 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
 
   where
     logger = scLogger serverCtx
-    encodeQErr' _ = encodeQErr
 
     spockAction
       :: (FromJSON a, ToJSON a, MonadIO m, UserAuthentication m, HttpLog m)
-      => (GraphQLResponseConfig -> Bool -> QErr -> Value)
+      => (Bool -> QErr -> Value)
       -> (QErr -> QErr) -> APIHandler m a -> Spock.ActionT m ()
     spockAction = mkSpockAction serverCtx
 
@@ -657,7 +655,7 @@ httpApp corsCfg serverCtx enableConsole consoleAssetsDir enableTelemetry = do
     allMod200 qe = qe { qeStatus = HTTP.status200 }
 
     gqlExplainAction =
-      spockAction (const encodeQErr) id $ mkPostHandler $
+      spockAction encodeQErr id $ mkPostHandler $
         mkAPIRespHandler gqlExplainHandler
 
     enableGraphQL = isGraphQLEnabled serverCtx
