+++ server/src-lib/Hasura/Server/Init.hs
@@ -1,5 +1,41 @@
+-- | Types and functions related to the server initialisation
 {-# LANGUAGE CPP #-}
-module Hasura.Server.Init where
+module Hasura.Server.Init
+  ( DbUid(..)
+  , getDbId
+
+  , PGVersion(..)
+  , getPgVersion
+
+  , InstanceId(..)
+  , generateInstanceId
+
+  , StartupTimeInfo(..)
+
+  -- * Server command related
+  , parseRawConnInfo
+  , mkRawConnInfo
+  , mkHGEOptions
+  , mkConnInfo
+  , serveOptionsParser
+
+  -- * Downgrade command related
+  , downgradeShortcuts
+  , downgradeOptionsParser
+
+  -- * Help footers
+  , mainCmdFooter
+  , serveCmdFooter
+
+  -- * Startup logs
+  , mkGenericStrLog
+  , mkGenericLog
+  , inconsistentMetadataLog
+  , serveOptsToLog
+  , connInfoToLog
+
+  , module Hasura.Server.Init.Config
+  ) where
 
 import qualified Data.Aeson                       as J
 import qualified Data.Aeson.Casing                as J
@@ -12,9 +48,7 @@ import qualified Database.PG.Query                as Q
 import qualified Language.Haskell.TH.Syntax       as TH
 import qualified Text.PrettyPrint.ANSI.Leijen     as PP
 
-import           Data.Char                        (toLower)
 import           Data.FileEmbed                   (embedStringFile)
-import           Data.Time.Clock.Units            (milliseconds)
 import           Network.Wai.Handler.Warp         (HostPreference)
 import           Options.Applicative
 
@@ -29,6 +63,7 @@ import           Hasura.RQL.Types                 (QErr, RoleName (..), SchemaCa
                                                    mkNonEmptyText)
 import           Hasura.Server.Auth
 import           Hasura.Server.Cors
+import           Hasura.Server.Init.Config
 import           Hasura.Server.Logging
 import           Hasura.Server.Utils
 import           Network.URI                      (parseURI)
@@ -63,203 +98,6 @@ data StartupTimeInfo
   , _stiTimeTaken :: !Double
   }
 $(J.deriveJSON (J.aesonDrop 4 J.snakeCase) ''StartupTimeInfo)
-
-data RawConnParams
-  = RawConnParams
-  { rcpStripes      :: !(Maybe Int)
-  , rcpConns        :: !(Maybe Int)
-  , rcpIdleTime     :: !(Maybe Int)
-  , rcpAllowPrepare :: !(Maybe Bool)
-  } deriving (Show, Eq)
-
-type RawAuthHook = AuthHookG (Maybe T.Text) (Maybe AuthHookType)
-
-data RawServeOptions impl
-  = RawServeOptions
-  { rsoPort                :: !(Maybe Int)
-  , rsoHost                :: !(Maybe HostPreference)
-  , rsoConnParams          :: !RawConnParams
-  , rsoTxIso               :: !(Maybe Q.TxIsolation)
-  , rsoAdminSecret         :: !(Maybe AdminSecret)
-  , rsoAuthHook            :: !RawAuthHook
-  , rsoJwtSecret           :: !(Maybe JWTConfig)
-  , rsoUnAuthRole          :: !(Maybe RoleName)
-  , rsoCorsConfig          :: !(Maybe CorsConfig)
-  , rsoEnableConsole       :: !Bool
-  , rsoConsoleAssetsDir    :: !(Maybe Text)
-  , rsoEnableTelemetry     :: !(Maybe Bool)
-  , rsoWsReadCookie        :: !Bool
-  , rsoStringifyNum        :: !Bool
-  , rsoEnabledAPIs         :: !(Maybe [API])
-  , rsoMxRefetchInt        :: !(Maybe LQ.RefetchInterval)
-  , rsoMxBatchSize         :: !(Maybe LQ.BatchSize)
-  , rsoEnableAllowlist     :: !Bool
-  , rsoEnabledLogTypes     :: !(Maybe [L.EngineLogType impl])
-  , rsoLogLevel            :: !(Maybe L.LogLevel)
-  , rsoPlanCacheSize       :: !(Maybe Cache.CacheSize)
-  , rsoDevMode             :: !Bool
-  , rsoAdminInternalErrors :: !Bool
-  }
-
-data ServeOptions impl
-  = ServeOptions
-  { soPort                :: !Int
-  , soHost                :: !HostPreference
-  , soConnParams          :: !Q.ConnParams
-  , soTxIso               :: !Q.TxIsolation
-  , soAdminSecret         :: !(Maybe AdminSecret)
-  , soAuthHook            :: !(Maybe AuthHook)
-  , soJwtSecret           :: !(Maybe JWTConfig)
-  , soUnAuthRole          :: !(Maybe RoleName)
-  , soCorsConfig          :: !CorsConfig
-  , soEnableConsole       :: !Bool
-  , soConsoleAssetsDir    :: !(Maybe Text)
-  , soEnableTelemetry     :: !Bool
-  , soStringifyNum        :: !Bool
-  , soEnabledAPIs         :: !(Set.HashSet API)
-  , soLiveQueryOpts       :: !LQ.LiveQueriesOptions
-  , soEnableAllowlist     :: !Bool
-  , soEnabledLogTypes     :: !(Set.HashSet (L.EngineLogType impl))
-  , soLogLevel            :: !L.LogLevel
-  , soPlanCacheOptions    :: !E.PlanCacheOptions
-  , soDevMode             :: !Bool
-  , soAdminInternalErrors :: !Bool
-  }
-
-data DowngradeOptions
-  = DowngradeOptions
-  { dgoTargetVersion :: !T.Text
-  , dgoDryRun        :: !Bool
-  } deriving (Show, Eq)
-
-data RawConnInfo =
-  RawConnInfo
-  { connHost     :: !(Maybe String)
-  , connPort     :: !(Maybe Int)
-  , connUser     :: !(Maybe String)
-  , connPassword :: !String
-  , connUrl      :: !(Maybe String)
-  , connDatabase :: !(Maybe String)
-  , connOptions  :: !(Maybe String)
-  , connRetries  :: !(Maybe Int)
-  } deriving (Eq, Read, Show)
-
-data HGECommandG a
-  = HCServe !a
-  | HCExport
-  | HCClean
-  | HCExecute
-  | HCVersion
-  | HCDowngrade !DowngradeOptions
-  deriving (Show, Eq)
-
-data API
-  = METADATA
-  | GRAPHQL
-  | PGDUMP
-  | DEVELOPER
-  | CONFIG
-  deriving (Show, Eq, Read, Generic)
-$(J.deriveJSON (J.defaultOptions { J.constructorTagModifier = map toLower })
-  ''API)
-
-instance Hashable API
-
-type HGECommand impl = HGECommandG (ServeOptions impl)
-type RawHGECommand impl = HGECommandG (RawServeOptions impl)
-
-data HGEOptionsG a
-  = HGEOptionsG
-  { hoConnInfo :: !RawConnInfo
-  , hoCommand  :: !(HGECommandG a)
-  } deriving (Show, Eq)
-
-type RawHGEOptions impl = HGEOptionsG (RawServeOptions impl)
-type HGEOptions impl = HGEOptionsG (ServeOptions impl)
-
-type Env = [(String, String)]
-
-class FromEnv a where
-  fromEnv :: String -> Either String a
-
-instance FromEnv String where
-  fromEnv = Right
-
-instance FromEnv HostPreference where
-  fromEnv = Right . DataString.fromString
-
-instance FromEnv Text where
-  fromEnv = Right . T.pack
-
-instance FromEnv AuthHookType where
-  fromEnv = readHookType
-
-instance FromEnv Int where
-  fromEnv = maybe (Left "Expecting Int value") Right . readMaybe
-
-instance FromEnv AdminSecret where
-  fromEnv = Right . AdminSecret . T.pack
-
-instance FromEnv RoleName where
-  fromEnv string = case mkNonEmptyText (T.pack string) of
-    Nothing     -> Left "empty string not allowed"
-    Just neText -> Right $ RoleName neText
-
-instance FromEnv Bool where
-  fromEnv = parseStrAsBool
-
-instance FromEnv Q.TxIsolation where
-  fromEnv = readIsoLevel
-
-instance FromEnv CorsConfig where
-  fromEnv = readCorsDomains
-
-instance FromEnv [API] where
-  fromEnv = readAPIs
-
-instance FromEnv LQ.BatchSize where
-  fromEnv = fmap LQ.BatchSize . readEither
-
-instance FromEnv LQ.RefetchInterval where
-  fromEnv = fmap (LQ.RefetchInterval . milliseconds . fromInteger) . readEither
-
-instance FromEnv JWTConfig where
-  fromEnv = readJson
-
-instance L.EnabledLogTypes impl => FromEnv [L.EngineLogType impl] where
-  fromEnv = L.parseEnabledLogTypes
-
-instance FromEnv L.LogLevel where
-  fromEnv = readLogLevel
-
-instance FromEnv Cache.CacheSize where
-  fromEnv = Cache.mkCacheSize
-
-parseStrAsBool :: String -> Either String Bool
-parseStrAsBool t
-  | map toLower t `elem` truthVals = Right True
-  | map toLower t `elem` falseVals = Right False
-  | otherwise = Left errMsg
-  where
-    truthVals = ["true", "t", "yes", "y"]
-    falseVals = ["false", "f", "no", "n"]
-
-    errMsg = " Not a valid boolean text. " ++ "True values are "
-             ++ show truthVals ++ " and  False values are " ++ show falseVals
-             ++ ". All values are case insensitive"
-
-readIsoLevel :: String -> Either String Q.TxIsolation
-readIsoLevel isoS =
-  case isoS of
-    "read-committed"  -> return Q.ReadCommitted
-    "repeatable-read" -> return Q.RepeatableRead
-    "serializable"    -> return Q.Serializable
-    _                 -> Left "Only expecting read-committed / repeatable-read / serializable"
-
-type WithEnv a = ReaderT Env (ExceptT String Identity) a
-
-runWithEnv :: Env -> WithEnv a -> Either String a
-runWithEnv env m = runIdentity $ runExceptT $ runReaderT m env
 
 returnJust :: Monad m => a -> m (Maybe a)
 returnJust = return . Just
@@ -350,13 +188,18 @@ mkServeOptions rso = do
   serverLogLevel <- fromMaybe L.LevelInfo <$> withEnv (rsoLogLevel rso) (fst logLevelEnv)
   planCacheOptions <- E.mkPlanCacheOptions <$> withEnv (rsoPlanCacheSize rso) (fst planCacheSizeEnv)
   devMode <- withEnvBool (rsoDevMode rso) $ fst devModeEnv
-  graphqlAdminInternalErrors <- withEnvBool (rsoAdminInternalErrors rso) $
-                                fst adminInternalErrorsEnv
+  adminInternalErrors <- fromMaybe True <$> -- Default to `true` to enable backwards compatibility
+                                withEnv (rsoAdminInternalErrors rso) (fst adminInternalErrorsEnv)
+  let internalErrorsConfig =
+        if | devMode             -> InternalErrorsAllRequests
+           | adminInternalErrors -> InternalErrorsAdminOnly
+           | otherwise           -> InternalErrorsDisabled
+
   return $ ServeOptions port host connParams txIso adminScrt authHook jwtSecret
                         unAuthRole corsCfg enableConsole consoleAssetsDir
                         enableTelemetry strfyNum enabledAPIs lqOpts enableAL
                         enabledLogs serverLogLevel planCacheOptions
-                        devMode graphqlAdminInternalErrors
+                        internalErrorsConfig
   where
 #ifdef DeveloperAPIs
     defaultAPIs = [METADATA,GRAPHQL,PGDUMP,CONFIG,DEVELOPER]
@@ -481,6 +324,9 @@ serveCmdFooter =
         ]
       , [ "# Start GraphQL Engine with HTTP compression enabled for '/v1/query' and '/v1/graphql' endpoints"
         , "graphql-engine --database-url <database-url> serve --enable-compression"
+        ]
+      , [ "# Start GraphQL Engine with enable/disable including 'internal' information in an error response for the request made by an 'admin'"
+        , "graphql-engine --database-url <database-url> serve --admin-internal-errors true|false"
         ]
       ]
 
@@ -671,7 +517,7 @@ devModeEnv =
 adminInternalErrorsEnv :: (String, String)
 adminInternalErrorsEnv =
   ( "HASURA_GRAPHQL_ADMIN_INTERNAL_ERRORS"
-  , "Include 'internal' key in the errors extensions (if required) of the response for GraphQL requests with the 'admin' role"
+  , "Enables including 'internal' information in an error response for requests made by an 'admin' (default: true)"
   )
 
 parseRawConnInfo :: Parser RawConnInfo
@@ -722,9 +568,6 @@ parseRawConnInfo =
                     metavar "NO OF RETRIES" <>
                     help (snd retriesNumEnv)
                   )
-
-connInfoErrModifier :: String -> String
-connInfoErrModifier s = "Fatal Error : " ++ s
 
 mkConnInfo :: RawConnInfo -> Either String Q.ConnInfo
 mkConnInfo (RawConnInfo mHost mPort mUser password mURL mDB opts mRetries) =
@@ -814,36 +657,6 @@ parseAdminSecret =
                 metavar "ADMIN SECRET KEY" <>
                 help (snd adminSecretEnv)
               )
-
-readHookType :: String -> Either String AuthHookType
-readHookType tyS =
-  case tyS of
-    "GET"  -> Right AHTGet
-    "POST" -> Right AHTPost
-    _      -> Left "Only expecting GET / POST"
-
-readAPIs :: String -> Either String [API]
-readAPIs = mapM readAPI . T.splitOn "," . T.pack
-  where readAPI si = case T.toUpper $ T.strip si of
-          "METADATA"  -> Right METADATA
-          "GRAPHQL"   -> Right GRAPHQL
-          "PGDUMP"    -> Right PGDUMP
-          "DEVELOPER" -> Right DEVELOPER
-          "CONFIG"    -> Right CONFIG
-          _            -> Left "Only expecting list of comma separated API types metadata,graphql,pgdump,developer,config"
-
-readLogLevel :: String -> Either String L.LogLevel
-readLogLevel s = case T.toLower $ T.strip $ T.pack s of
-  "debug" -> Right L.LevelDebug
-  "info"  -> Right L.LevelInfo
-  "warn"  -> Right L.LevelWarn
-  "error" -> Right L.LevelError
-  _       -> Left "Valid log levels: debug, info, warn or error"
-
-
-readJson :: (J.FromJSON a) => String -> Either String a
-readJson = J.eitherDecodeStrict . txtToBs . T.pack
-
 
 parseWebHook :: Parser RawAuthHook
 parseWebHook =
@@ -971,9 +784,10 @@ parseGraphqlDevMode =
            help (snd devModeEnv)
          )
 
-parseGraphqlAdminInternalErrors :: Parser Bool
-parseGraphqlAdminInternalErrors =
-  switch ( long "admin-internal-errors" <>
+parseGraphqlAdminInternalErrors :: Parser (Maybe Bool)
+parseGraphqlAdminInternalErrors = optional $
+  option (eitherReader parseStrAsBool)
+         ( long "admin-internal-errors" <>
            help (snd adminInternalErrorsEnv)
          )
