+++ server/src-lib/Hasura/Server/SchemaUpdate.hs
@@ -5,13 +5,13 @@ where
 import           Hasura.Prelude
 
 import           Hasura.Logging
-import           Hasura.RQL.DDL.Schema     (runCacheRWT)
+import           Hasura.RQL.DDL.Schema       (runCacheRWT)
 import           Hasura.RQL.Types
 import           Hasura.RQL.Types.Run
-import           Hasura.Server.App         (SchemaCacheRef (..), withSCUpdate)
-import           Hasura.Server.Init        (InstanceId (..))
+import           Hasura.Server.API.Query
+import           Hasura.Server.App           (SchemaCacheRef (..), withSCUpdate)
+import           Hasura.Server.Init          (InstanceId (..))
 import           Hasura.Server.Logging
-import           Hasura.Server.Query
 
 import           Data.Aeson
 import           Data.Aeson.Casing
@@ -93,7 +93,7 @@ startSchemaSyncThreads sqlGenCtx pool logger httpMgr cacheRef instanceId cacheIn
   updateEventRef <- liftIO $ STM.newTVarIO Nothing
 
   -- Start listener thread
-  lTId <- liftIO $ C.forkImmortal "SchemeUpdate.listener" logger $ 
+  lTId <- liftIO $ C.forkImmortal "SchemeUpdate.listener" logger $
     listener sqlGenCtx pool logger httpMgr updateEventRef cacheRef instanceId cacheInitTime
   logThreadStarted TTListener lTId
 
@@ -165,7 +165,7 @@ listener sqlGenCtx pool logger httpMgr updateEventRef
             STM.atomically $ STM.writeTVar updateEventRef $ Just payload
 
     onError = logError logger threadType . TEQueryError
-    -- NOTE: we handle expected error conditions here, while unexpected exceptions will result in 
+    -- NOTE: we handle expected error conditions here, while unexpected exceptions will result in
     -- a restart and log from 'forkImmortal'
     logWarn = unLogger logger $
       SchemaSyncThreadLog LevelWarn TTListener $ String
