+++ server/tests-py/queries/actions/sync/expecting_array_response.yaml
@@ -37,29 +37,27 @@
     errors:
     - extensions:
         internal:
-          debug_info:
-            response:
-              status: 200
-              body:
-                id: 1
-              headers:
-              - value: application/json
-                name: Content-Type
-              - value: abcd
-                name: Set-Cookie
-            request:
-              body:
-                session_variables:
-                  x-hasura-role: admin
-                input:
-                  email: clarke@gmail.com
-                  name: Clarke
-                action:
-                  name: create_users
-              url: http://127.0.0.1:5593/create-user
-              headers: []
-          error: expecting array for action webhook response but got object
-          type: unexpected_json
+          error: unexpected response
+          response:
+            status: 200
+            body:
+              id: 1
+            headers:
+            - value: application/json
+              name: Content-Type
+            - value: abcd
+              name: Set-Cookie
+          request:
+            body:
+              session_variables:
+                x-hasura-role: admin
+              input:
+                email: clarke@gmail.com
+                name: Clarke
+              action:
+                name: create_users
+            url: http://127.0.0.1:5593/create-user
+            headers: []
         path: $
         code: unexpected
       message: expecting array for action webhook response but got object
