+++ server/tests-py/queries/actions/sync/expecting_object_response.yaml
@@ -35,33 +35,31 @@
     errors:
     - extensions:
         internal:
-          debug_info:
-            response:
-              status: 200
-              body:
-              - id: 1
-              - id: 2
-              headers:
-              - value: application/json
-                name: Content-Type
-              - value: abcd
-                name: Set-Cookie
-            request:
-              body:
-                session_variables:
-                  x-hasura-role: admin
-                input:
-                  users:
-                  - email: blake@hasura.io
-                    name: Blake
-                  - email: elsa@hasura.io
-                    name: Elsa
-                action:
-                  name: create_user
-              url: http://127.0.0.1:5593/create-users
-              headers: []
-          error: expecting object for action webhook response but got array
-          type: unexpected_json
+          error: unexpected response
+          response:
+            status: 200
+            body:
+            - id: 1
+            - id: 2
+            headers:
+            - value: application/json
+              name: Content-Type
+            - value: abcd
+              name: Set-Cookie
+          request:
+            body:
+              session_variables:
+                x-hasura-role: admin
+              input:
+                users:
+                - email: blake@hasura.io
+                  name: Blake
+                - email: elsa@hasura.io
+                  name: Elsa
+              action:
+                name: create_user
+            url: http://127.0.0.1:5593/create-users
+            headers: []
         path: $
         code: unexpected
       message: expecting object for action webhook response but got array
