+++ server/tests-py/queries/actions/sync/invalid_webhook_response.yaml
@@ -38,28 +38,26 @@
     errors:
     - extensions:
         internal:
-          debug_info:
-            response:
-              status: 200
-              body: some-string
-              headers:
-              - value: application/json
-                name: Content-Type
-              - value: abcd
-                name: Set-Cookie
-            request:
-              body:
-                session_variables:
-                  x-hasura-role: admin
-                input:
-                  email: clarke@gmail.com
-                  name: Clarke
-                action:
-                  name: create_user
-              url: http://127.0.0.1:5593/invalid-response
-              headers: []
-          error: expecting object or array of objects for action webhook response
-          type: unexpected_json
+          error: unexpected response
+          response:
+            status: 200
+            body: some-string
+            headers:
+            - value: application/json
+              name: Content-Type
+            - value: abcd
+              name: Set-Cookie
+          request:
+            body:
+              session_variables:
+                x-hasura-role: admin
+              input:
+                email: clarke@gmail.com
+                name: Clarke
+              action:
+                name: create_user
+            url: http://127.0.0.1:5593/invalid-response
+            headers: []
         path: $
         code: parse-failed
       message: expecting object or array of objects for action webhook response
