+++ server/tests-py/queries/actions/sync/mirror_action_no_field.yaml
@@ -5,29 +5,27 @@ response:
   errors:
   - extensions:
       internal:
-        debug_info:
-          response:
-            status: 200
-            body:
-              name: Alice
-            headers:
-            - value: application/json
-              name: Content-Type
-            - value: abcd
-              name: Set-Cookie
-          request:
-            body:
-              session_variables:
-                x-hasura-role: admin
-              input:
-                arg:
-                  name: Alice
-              action:
-                name: mirror
-            url: http://127.0.0.1:5593/mirror-action
-            headers: []
-        error: field "id" expected in webhook response, but not found
-        type: unexpected_json
+        error: unexpected response
+        response:
+          status: 200
+          body:
+            name: Alice
+          headers:
+          - value: application/json
+            name: Content-Type
+          - value: abcd
+            name: Set-Cookie
+        request:
+          body:
+            session_variables:
+              x-hasura-role: admin
+            input:
+              arg:
+                name: Alice
+            action:
+              name: mirror
+          url: http://127.0.0.1:5593/mirror-action
+          headers: []
       path: $
       code: unexpected
     message: field "id" expected in webhook response, but not found
