+++ server/tests-py/validate.py
@@ -331,11 +331,6 @@ def equal_CommentedMap(m1, m2):
     else:
         return m1 == m2
 
-# Parse test case YAML file
-def get_conf_f(f):
-    with open(f, 'r+') as c:
-        return yaml.YAML().load(c)
-
 def check_query_f(hge_ctx, f, transport='http', add_auth=True):
     print("Test file: " + f)
     hge_ctx.may_skip_test_teardown = False
